Rename HookConfiguration to CaptainHook::Configuration
Follow the starndard ruby gem folder structure with a gem_name.rb
and a gem_name folder with the rest of the library

This makes it predictable for external developers and it will prevent
from polluting the global namespace with classes from the gem
